:crystal_ball: Clairvoyance and clarity.

+++ conjure.sh
@@ -5,13 +5,12 @@
 # ------------------ ---------  ------   ----    ---     --      -
 #%
 #% DESCRIPTION
-#%    Magical Wordpress development environment
-#%    Based on Laravel Homestead, modified to include
-#%    plenty of wordpress specific tools and optimizations
-#%    for development and deployment.
+#% ===========
+#%    Magical Wordpress development environment, Based on Laravel Homestead!
+#%    Modified to include plenty of Wordpress tools and optimizations
 #%
 #+ COMMANDS
-#+ -----------------------------------------------------------------------------------------------------------
+#+ ===========
 #+    setup     . . . . . . . .  Installs WP_Conjure core system files
 #+    wpnew     . . . . . . . .  Installs new Wordpress site into WP_Conjure directory: www
 #+
@@ -20,16 +19,11 @@
 #+
 #+    stacks    . . . . . . . .  Display a list of alternative Wordpress stacks to try out
 #+    stackadd  . . . . . . . .  Downloads a new Wordpress for a given webstack. [plate,rock,web]
-#+     ------                     ------      ---    ------
-#+    -> plate  . . . . . . . .  Wordplate | ***  | https://wordplate.github.io/"
-#+    ->  rock  . . . . . . . .  Bedrock   | **** | https://roots.io/bedrock/"
-#+    ->   web  . . . . . . . .  Webpress  | **   | https://bitbucket.org/derekscott_mm/webpress"
-#+    -> plain  . . . . . . . .  Wordpress |      | https://github.com/johnpbloch/wordpress-core-installer"
 #+
-#%
 #% INFO
-#%    -h, --help                    Print this help
-#%    -v, --version                 Print script information
+#% ===========
+#%    -h, --help    . . . . . .  Print this help
+#%    -v, --version . . . . . .  Print script information
 #%
 # ------------------ ---------  ------   ----    ---     --      -
 ## TITLE:    init.sh
@@ -47,8 +41,8 @@ SCRIPTNAME=$(basename $0)
 SCRIPTBASENAME="$(basename ${SCRIPTNAME} .sh)"
 SCRIPT_HEADSIZE=$(head -200 ${0} | grep -n "^# END_OF_HEADER" | cut -f1 -d:)
 SCRIPT_NAME="$(basename ${0})"
-vVPress="0.7.0";
-scriptPath="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+vVPress="0.7.0"
+scriptPath="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 scriptName=$(basename $0)                      # Set Script Name variable
 scriptBasename="$(basename ${scriptName} .sh)" # Strips '.sh' from scriptName
 args=()
@@ -70,75 +64,78 @@ source $installScript
 # Main Menu
 # mageHatLogo
 
-function installhs {
+function installhs() {
   echo "Running Install script..."
 
-  if [ -f "./.conjure.json"]; then
+  if [ -f "./.conjure.json" ]; then
     mv ./.conjure.json ./conjure-old.json
   fi
 
-  if [ -f "./.env"]; then
+  if [ -f "./.env" ]; then
     mv ./.env ./env-old.txt
   fi
 
-  rm -R ./cache/*
-  rm -R ./log/*.log
+  #rm -R ./cache
+  if [ -f "./log/*.log" ]; then
+    rm -R ./log/*.log
+  fi
 
-  RUNINSTALLER
+  conjureFileRoll
 }
 
-function listStacks {
-  echo "${NORMAL}${GRN}                  Available Wordpress Stacks ${NORMAL}"
+function listStacks() {
+  GRAY="$(tput setaf 244)"
+  echo "${NORMAL}
+ Wordpress Stacks
+ ==========="
   echo ""
-  echo "${CYN}  Slug                   Fullname   DIF    URL "
-  echo "${BBLK}  -----------------------------------------------------------"
+  echo "${BGRN}  Slug                   Fullname   DIF    URL "
+  echo "${GRAY}  -----------------------------------------------------------"
+  echo "${BBLU}  plate ${NORMAL}. . . . . . . .  ${BBLU}Wordplate ${GRAY}| ${YLW}xxx  ${GRAY}| ${CYN}https://wordplate.github.io/"
+  echo "${BBLU}  rock  ${NORMAL}. . . . . . . .  ${BBLU}Bedrock   ${GRAY}| ${YLW}xxxx ${GRAY}| ${CYN}https://roots.io/bedrock/"
+  echo "${BBLU}  web   ${NORMAL}. . . . . . . .  ${BBLU}Webpress  ${GRAY}| ${YLW}xx   ${GRAY}| ${CYN}https://bitbucket.org/derekscott_mm/webpress"
+  echo "${BBLU}  plain ${NORMAL}. . . . . . . .  ${BBLU}Wordpress ${GRAY}|      ${GRAY}| ${CYN}https://github.com/johnpbloch/wordpress-core-installer"
   echo ""
-  echo "${BBLU}  plate ${BBLK}. . . . . . . .  ${BLU}Wordplate ${BBLK}| ${YLW}***  ${BBLK}| ${CYN}https://wordplate.github.io/"
-  echo "${BBLU}  rock  ${BBLK}. . . . . . . .  ${BLU}Bedrock   ${BBLK}| ${YLW}**** ${BBLK}| ${CYN}https://roots.io/bedrock/"
-  echo "${BBLU}  web   ${BBLK}. . . . . . . .  ${BLU}Webpress  ${BBLK}| ${YLW}**   ${BBLK}| ${CYN}https://bitbucket.org/derekscott_mm/webpress"
-  echo "${BBLU}  plain ${BBLK}. . . . . . . .  ${BLU}Wordpress ${BBLK}|      ${BBLK}| ${CYN}https://github.com/johnpbloch/wordpress-core-installer"
   echo ""
-  echo ""
-  echo "${PUR}  * [DIF]FERENTIATE = How different that stack is compared to default Wordpress"
+  echo "${YLW}  * [DIF]FERENTIATE = How different that stack is compared to default Wordpress"
   echo "${NORMAL}"
   safeExit
 }
 
-function newStack {
+function newStack() {
   local item
-  item="${args[1]}";
+  item="${args[1]}"
   case $item in
-    plate) addStackPlate ;;
-    rock) addStackRock ;;
-    web) addStackWeb ;;
-    plain) addStackOrig ;;
+  plate) addStackPlate ;;
+  rock) addStackRock ;;
+  web) addStackWeb ;;
+  plain) addStackOrig ;;
   esac
 }
 
-function addStackPlate {
+function addStackPlate() {
   echo "Adding Wordplate site..."
   cd ./CLI/wp_build && bash Wordplate.sh
 }
 
-function addStackRock {
+function addStackRock() {
   echo "Adding Bedrock site..."
   new_bedrock
 }
 
-function addStackOrig {
+function addStackOrig() {
   echo "Adding Wordpress site..."
   GETNEWWP
 }
 
-function addStackWeb {
+function addStackWeb() {
   echo "Adding WebPress site..."
   new_webpress
 }
 
-function vagrantCMD {
+function vagrantCMD() {
   echo "$@"
 }
-
 
 # Print help if no arguments were passed.
 # Uncomment to force arguments when invoking the script
@@ -157,36 +154,44 @@ function mainInitScript() {
   GETTHISPARTYSTARTED
 
   # Read the options and set stuff
-  while [[ $1 = -?* ]]; do
+  while [[ $1 == -?* ]]; do
     case $1 in
-      -h|--help) usagefull >&2; safeExit ;;
-      -v|--version) echo "WP_Conjure | version ${CONJUREVersion}"; safeExit ;;
-      -l|--log) printLog=1 ;;
-      -d|--debug) debug=1;;
-      --force) force=1 ;;
-      --endopts) shift; break ;;
+    -h | --help)
+      echo "${NORMAL}" && usagefull && listStacks >&2
+      safeExit
+      ;;
+    -v | --version)
+      echo "WP_Conjure | version ${CONJUREVersion}"
+      safeExit
+      ;;
+    -l | --log) printLog=1 ;;
+    -d | --debug) debug=1 ;;
+    --force) force=1 ;;
+    --endopts)
+      shift
+      break
+      ;;
     esac
     shift
   done
 
   case $1 in
-    setup) installhs ;;
-    wpnew) GETNEWWP ;;
-    vagrant) vagrantCMD $args ;;
-    wordmove) newMoveFile ;;
-    stacks) listStacks ;;
-    stackadd) newStack $args ;;
+  setup) installhs ;;
+  wpnew) GETNEWWP ;;
+  vagrant) vagrantCMD $args ;;
+  wordmove) newMoveFile ;;
+  stacks) listStacks ;;
+  stackadd) newStack $args ;;
   esac
 
   # If no params (argument array 0 is null)
   if [ -z ${args[0]} ]; then
     echo
-    echo "${BGRN}Heyo there buddy${NORMAL}. You didnt provide any arguments so I'm not doing anything."
-    echo
-    usage
+    echo "${BGRN}Try harder next time.${NORMAL}. You didnt provide any arguments, so I'm doing nothing..."
+    echo "${NORMAL}"
+    #usage
     safeExit
   fi
-
 
   ####################################################
   ############### End Script Here ####################
@@ -225,16 +230,7 @@ while (($#)); do
 done
 set -- "${options[@]}"
 
-
 # Store the remaining part as arguments.
 args+=("$@")
 
 mainInitScript $@
-
-
-function info { echo "${BOLD}${WHT}${B_BLU}[INFO]${NORMAL}${BLU} ➜${NORMAL} $@" }
-function warning { echo "${BOLD}${WHT}${B_GRN}[ OK ]${NORMAL}${GRN} ✔${NORMAL} $@" }
-function warning { echo "${BOLD}${WHT}${B_YLW}[WARN]${NORMAL}${YLW} ➜${NORMAL} $@" }
-function error { echo "${BOLD}${WHT}${B_RED}[FAIL]${NORMAL}${RED} ✖${NORMAL} $@" }
-function fatal { echo "${BOLD}${WHT}${B_RED}[EXIT]${NORMAL}${RED} ✖${NORMAL} $@" && exit 0 }
-function die { echo "${BOLD}${WHT}${B_RED}[DIE!]${NORMAL}${RED} ✖${NORMAL} $@" && exit 1 }
